backend/requirements.txt
pinecone-client 
cohere 
flask 
python-dotenv